+++ src/ble/reducers.ts
@@ -1,11 +1,13 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2020 The Pybricks Authors
+// Copyright (c) 2020-2021 The Pybricks Authors
 //
 // Manages state for the Bluetooth Low Energy connection.
 // This assumes that there is only one global connection to a single device.
 
 import { Reducer, combineReducers } from 'redux';
 import { Action } from '../actions';
+import { BleDIServiceActionType } from '../ble-device-info-service/actions';
+import { getHubTypeName } from '../ble-device-info-service/protocol';
 import { BleDeviceActionType } from './actions';
 
 /**
@@ -50,4 +52,42 @@ const connection: Reducer<BleConnectionState, Action> = (
     }
 };
 
-export default combineReducers({ connection });
+const deviceName: Reducer<string, Action> = (state = '', action) => {
+    switch (action.type) {
+        case BleDeviceActionType.DidDisconnect:
+            return '';
+        case BleDeviceActionType.DidConnect:
+            return action.name;
+        default:
+            return state;
+    }
+};
+
+const deviceType: Reducer<string, Action> = (state = '', action) => {
+    switch (action.type) {
+        case BleDeviceActionType.DidDisconnect:
+            return '';
+        case BleDIServiceActionType.DidReceivePnPId:
+            return getHubTypeName(action.pnpId);
+        default:
+            return state;
+    }
+};
+
+const deviceFirmwareVersion: Reducer<string, Action> = (state = '', action) => {
+    switch (action.type) {
+        case BleDeviceActionType.DidDisconnect:
+            return '';
+        case BleDIServiceActionType.DidReceiveFirmwareRevision:
+            return action.version;
+        default:
+            return state;
+    }
+};
+
+export default combineReducers({
+    connection,
+    deviceName,
+    deviceType,
+    deviceFirmwareVersion,
+});