+++ src/status-bar/i18n.ts
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2021 The Pybricks Authors
+//
+// Status bar translation keys.
+
+export enum MessageId {
+    HubInfoConnectedTo = 'hubInfo.connectedTo',
+    HubInfoHubType = 'hubInfo.hubType',
+    HubInfoFirmware = 'hubInfo.firmware',
+}